Document design decisions (#14)
We document some of the design decisions. The majority of the text comes
from the other SDKs, and we simply adapted it to Python.

We also document why runtime type guards are not used.
Fixes #436.

Additionally, we make some minor fixes to changelog so that the docs
compile without warnings (no content is affected, just the indention).