Test and fix for text attached to end XML elements
The ElementTree's `XMLPullParser` sometimes attaches the text to the
start element and sometimes to the end element. We could not figure out
why there is a different behavior, but suspect that it has something to
do with the size of the parser buffer.

In this patch, we fix by looking for text in both events (start as well
as end element).

This is related to the issue #17, and based on the aas-core-codegen
after the pull request
https://github.com/aas-core-works/aas-core-codegen/pull/443 has been
merged in, [aas-core-codegen 330f391].

[aas-core-codegen 330f391]: https://github.com/aas-core-works/aas-core-codegen/commit/330f391