Check explicitly for missing model type
We add an explicit check for missing ``modelType`` to account for
the case where a concrete class without descendants needs a model type
for backwards compatibility even though the model type property is
redundant.

The code corresponds to [aas-core-codegen 84ccbb64], and the test data
correspods to [aas-core3.0-testgen 1b00dbe45].

[aas-core-codegen 84ccbb64]: https://github.com/aas-core-works/aas-core-codegen/commit/84ccbb64
[aas-core3.0-testgen 1b00dbe45]: https://github.com/aas-core-works/aas-core3.0-testgen/commit/1b00dbe45

Fixes #32.