Remove checks from update script (#54)
We do not run the checks in the update script since we want to make it
compatible with other Python versions, not just Python 3.10. In
particular, pylint changes behavior from Python version to Python
version -- and this we want to avoid.

Instead, we let the remote CI do the checks using a single Python 3.10.